guard all mex calls against single precision input
The mex interfaces all require double: the nfft/nfsft/nfsoft ones reject
single with an error, while the C mex files reading their input with
mxGetPr / mxGetDoubles (S1Grid_find, S2Grid_find, SO3Grid_find,
EulerCyclesC, ...) would silently misinterpret or crash on it. Cast the
numeric arguments to double at every call site, so single precision data
- e.g. an EBSD map imported from a file storing 4 byte floats - can no
longer take down a transform or a grid search.

Same for sparse() in SO3Kernel/K_symmetrised, which does not accept
single precision values either.

+++ EBSDAnalysis/@EBSD/private/spatialDecompositionGrid.m
@@ -281,9 +281,9 @@ site2id  = [idxSiteEbsd; niId];
 % sizing-only pass.
 if check_option(varargin,'delaunayOnly')
   V = zeros(0,2); F = zeros(0,2);
-  I_FD = jcvoronoiDelaunayOnly_mex(double(XY),double(numReal), epsilon);
+  I_FD = jcvoronoiDelaunayOnly_mex(double(XY),double(numReal), double(epsilon));
 else
-  [V,F,I_FD] = jcvoronoi2_mex(double(XY),double(numReal), epsilon);
+  [V,F,I_FD] = jcvoronoi2_mex(double(XY),double(numReal), double(epsilon));
 end
 
 out = struct('V',V,'F',F,'I_FD',I_FD, ...
